object: verify in-db attributes
The initial db for the tests are created from ptool, so verify that
all the rsa and ecc attrs are as expected.

Move code from test.h into it's own test lib.
