+++ test/integration/pkcs-keygen.int.c
@@ -7,9 +7,7 @@
 
 #include <openssl/evp.h>
 #include <openssl/rsa.h>
-#include <openssl/bn.h>
 #include <openssl/err.h>
-#include <openssl/objects.h>
 
 #include "test.h"
 
@@ -57,249 +55,6 @@ static int test_teardown(void **state) {
     free(ti);
 
     return 0;
-}
-
-#define ADD_ATTR_BASE(t, x)  { .type = t,   .ulValueLen = sizeof(x),     .pValue = &x }
-#define ADD_ATTR_ARRAY(t, x) { .type = t,   .ulValueLen = ARRAY_LEN(x),  .pValue = x }
-#define ADD_ATTR_STR(t, x)   { .type = t,   .ulValueLen = sizeof(x) - 1, .pValue = x }
-
-GENERIC_ATTR_TYPE_CONVERT(CK_BBOOL);
-GENERIC_ATTR_TYPE_CONVERT(CK_ULONG);
-
-static void verify_missing_common_attrs_rsa(CK_SESSION_HANDLE session, CK_OBJECT_HANDLE h) {
-
-    CK_BYTE tmp[3][256] = { 0 };
-
-    CK_ATTRIBUTE attrs[] = {
-            ADD_ATTR_ARRAY(CKA_MODULUS,  tmp[0]),
-            ADD_ATTR_ARRAY(CKA_MODULUS_BITS,  tmp[1]),
-            ADD_ATTR_ARRAY(CKA_PUBLIC_EXPONENT, tmp[2]),
-    };
-
-    CK_RV rv = C_GetAttributeValue(session, h, attrs, ARRAY_LEN(attrs));
-    assert_int_equal(rv, CKR_OK);
-
-    CK_ULONG count = 0;
-
-    CK_ULONG i;
-    for (i=0; i < ARRAY_LEN(attrs); i++) {
-        CK_ATTRIBUTE_PTR a = &attrs[i];
-        switch(a->type) {
-        case CKA_MODULUS: {
-            assert_int_not_equal(0, a->ulValueLen);
-            assert_non_null(a->pValue);
-            count++;
-        } break;
-        case CKA_MODULUS_BITS: {
-            CK_ULONG v = 0;
-            rv = generic_CK_ULONG(a, &v);
-            assert_int_equal(rv, CKR_OK);
-            assert_int_equal(v, 2048);
-            count++;
-        } break;
-        case CKA_PUBLIC_EXPONENT:
-            assert_int_not_equal(0, a->ulValueLen);
-            assert_non_null(a->pValue);
-            count++;
-            break;
-        default:
-            assert_true(0);
-        }
-    }
-
-    assert_int_equal(count, ARRAY_LEN(attrs));
-}
-
-static void verify_missing_priv_attrs_rsa(CK_SESSION_HANDLE session, CK_OBJECT_HANDLE h) {
-    verify_missing_common_attrs_rsa(session, h);
-}
-
-static void verify_missing_pub_attrs_rsa(CK_SESSION_HANDLE session, CK_OBJECT_HANDLE h) {
-    verify_missing_common_attrs_rsa(session, h);
-}
-
-static void test_ec_params(CK_ATTRIBUTE_PTR ecparams) {
-
-    const unsigned char *p = ecparams->pValue;
-
-    ASN1_OBJECT *a = d2i_ASN1_OBJECT(NULL, &p, ecparams->ulValueLen);
-    assert_non_null(a);
-
-    int nid = OBJ_obj2nid(a);
-    ASN1_OBJECT_free(a);
-
-    switch (nid) {
-    case NID_X9_62_prime192v1:
-    case NID_secp224r1:
-    case NID_X9_62_prime256v1:
-    case NID_secp384r1:
-    case NID_secp521r1:
-        break;
-    default:
-        fail_msg("Unsupported nid to tpm EC algorithm mapping, got nid: %d", nid);
-    }
-}
-
-static void test_ec_point(CK_ATTRIBUTE_PTR ecpoint) {
-
-    assert_int_not_equal(0, ecpoint->ulValueLen);
-    assert_non_null(ecpoint->pValue);
-
-    const unsigned char *pp = ecpoint->pValue;
-    CK_ULONG len = ecpoint->ulValueLen;
-
-    ASN1_OCTET_STRING *a = NULL;
-    a = d2i_ASN1_OCTET_STRING(&a, &pp, len);
-    assert_non_null(a);
-
-    const unsigned char *d = ASN1_STRING_get0_data(a);
-
-    /* first byte should be 04 for "uncompressed format" */
-    assert_int_equal(d[0], 0x4);
-
-    /* TODO look at curve id and map to expected X and Y sizes */
-
-    ASN1_STRING_free(a);
-}
-
-static void verify_missing_pub_attrs_ecc(CK_SESSION_HANDLE session, CK_OBJECT_HANDLE h) {
-
-    CK_BYTE tmp[2][256] = { 0 };
-
-    CK_ATTRIBUTE attrs[] = {
-            ADD_ATTR_ARRAY(CKA_EC_PARAMS,  tmp[0]),
-            ADD_ATTR_ARRAY(CKA_EC_POINT,    tmp[1]),
-    };
-
-    CK_RV rv = C_GetAttributeValue(session, h, attrs, ARRAY_LEN(attrs));
-    assert_int_equal(rv, CKR_OK);
-
-    CK_ULONG count = 0;
-    CK_ULONG i;
-    for (i=0; i < ARRAY_LEN(attrs); i++) {
-        CK_ATTRIBUTE_PTR a = &attrs[i];
-        switch(a->type) {
-        /* TODO more robust checking here:
-         *  - They match what was expected in generation.
-         */
-        case CKA_EC_PARAMS:
-            test_ec_params(a);
-            count++;
-            break;
-        case CKA_EC_POINT:
-            // DER-encoding of ANSI X9.62 ECPoint value Q
-            test_ec_point(a);
-            count++;
-        break;
-        default:
-            assert_true(0);
-        }
-    }
-
-    assert_int_equal(count, ARRAY_LEN(attrs));
-}
-
-static void verify_missing_priv_attrs_common(CK_SESSION_HANDLE session, CK_KEY_TYPE keytype, CK_OBJECT_HANDLE h) {
-
-    CK_BYTE tmp[5][256] = { 0 };
-
-    CK_ATTRIBUTE attrs[] = {
-            ADD_ATTR_ARRAY(CKA_KEY_TYPE, tmp[0]),
-            ADD_ATTR_ARRAY(CKA_CLASS, tmp[1]),
-            ADD_ATTR_ARRAY(CKA_ALWAYS_SENSITIVE,  tmp[2]),
-            ADD_ATTR_ARRAY(CKA_EXTRACTABLE,  tmp[3]),
-            ADD_ATTR_ARRAY(CKA_NEVER_EXTRACTABLE,  tmp[4]),
-    };
-
-    CK_RV rv = C_GetAttributeValue(session, h, attrs, ARRAY_LEN(attrs));
-    assert_int_equal(rv, CKR_OK);
-
-    CK_ULONG count = 0;
-
-    CK_ULONG i;
-    for (i=0; i < ARRAY_LEN(attrs); i++) {
-        CK_ATTRIBUTE_PTR a = &attrs[i];
-        switch(a->type) {
-        case CKA_KEY_TYPE: {
-            CK_ULONG v = 0;
-            rv = generic_CK_ULONG(a, &v);
-            assert_int_equal(rv, CKR_OK);
-            assert_int_equal(v, keytype);
-            count++;
-        } break;
-        case CKA_CLASS: {
-            CK_ULONG v = 0;
-            rv = generic_CK_ULONG(a, &v);
-            assert_int_equal(rv, CKR_OK);
-            assert_int_equal(v, CKO_PRIVATE_KEY);
-            count++;
-        } break;
-        case CKA_ALWAYS_SENSITIVE: {
-            CK_BBOOL v = CK_FALSE;
-            rv = generic_CK_BBOOL(a, &v);
-            assert_int_equal(rv, CKR_OK);
-            assert_int_equal(v, CK_TRUE);
-            count++;
-        } break;
-        case CKA_EXTRACTABLE: {
-            CK_BBOOL v = CK_TRUE;
-            rv = generic_CK_BBOOL(a, &v);
-            assert_int_equal(rv, CKR_OK);
-            assert_int_equal(v, CK_FALSE);
-            count++;
-        } break;
-        case CKA_NEVER_EXTRACTABLE: {
-            CK_BBOOL v = CK_FALSE;
-            rv = generic_CK_BBOOL(a, &v);
-            assert_int_equal(rv, CKR_OK);
-            assert_int_equal(v, CK_TRUE);
-            count++;
-        } break;
-        default:
-            assert_true(0);
-        }
-    }
-
-    assert_int_equal(count, ARRAY_LEN(attrs));
-}
-
-static void verify_missing_pub_attrs_common(CK_SESSION_HANDLE session, CK_KEY_TYPE keytype, CK_OBJECT_HANDLE h) {
-
-    CK_BYTE tmp[2][256] = { 0 };
-
-    CK_ATTRIBUTE attrs[] = {
-            ADD_ATTR_ARRAY(CKA_KEY_TYPE, tmp[0]),
-            ADD_ATTR_ARRAY(CKA_CLASS,    tmp[1]),
-    };
-
-    CK_RV rv = C_GetAttributeValue(session, h, attrs, ARRAY_LEN(attrs));
-    assert_int_equal(rv, CKR_OK);
-
-    CK_ULONG count = 0;
-    CK_ULONG i;
-    for (i=0; i < ARRAY_LEN(attrs); i++) {
-        CK_ATTRIBUTE_PTR a = &attrs[i];
-        switch(a->type) {
-        case CKA_KEY_TYPE: {
-            CK_ULONG v = 0;
-            rv = generic_CK_ULONG(a, &v);
-            assert_int_equal(rv, CKR_OK);
-            assert_int_equal(v, keytype);
-            count++;
-        } break;
-        case CKA_CLASS: {
-            CK_ULONG v = 0;
-            rv = generic_CK_ULONG(a, &v);
-            assert_int_equal(rv, CKR_OK);
-            assert_int_equal(v, CKO_PUBLIC_KEY);
-            count++;
-        } break;
-        default:
-            assert_true(0);
-        }
-    }
-
-    assert_int_equal(count, ARRAY_LEN(attrs));
 }
 
 static void test_rsa_keygen_p11tool_templ(void **state) {
@@ -532,7 +287,9 @@ static void test_ecc_keygen_p11tool_templ(void **state) {
     /* verify missing attrs */
     verify_missing_pub_attrs_common(session, CKK_EC, pub_handle_dup);
     verify_missing_priv_attrs_common(session, CKK_EC, priv_handle_dup);
+
     verify_missing_pub_attrs_ecc(session, pub_handle_dup);
+    verify_missing_priv_attrs_ecc(session, priv_handle_dup);
 }
 
 static void test_destroy(void **state) {
